+++ hexital/exceptions.py
@@ -18,11 +18,6 @@ class InvalidCandleOrder(Exception):
         super().__init__(message)
 
 
-class CandleAlreadyTagged(Exception):
-    def __init__(self, message):
-        super().__init__(message)
-
-
 class InvalidCandlestickType(Exception):
     def __init__(self, message):
         super().__init__(message)