+++ tests/candlesticks/test_heikinashi.py
@@ -7,7 +7,7 @@ from hexital.core.candle import Candle
 
 @pytest.mark.usefixtures("candles", "candles_heikinashi")
 def test_heikinashi(candles: List[Candle], candles_heikinashi: List[Candle]):
-    heikin_ashi = HeikinAshi()
-    heikin_ashi.conversion(candles)
+    heikin_ashi = HeikinAshi(candles)
+    heikin_ashi.transform()
 
-    assert candles == candles_heikinashi
+    assert heikin_ashi.derived_candles == candles_heikinashi