+++ CHANGELOG.md
@@ -15,12 +15,15 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
     - Renamed Indicator 'as_list' to 'readings'
     - Renamed 'find_candles' to 'find_candle_pairing'
     - Removed 'has_reading' as pointless with method 'exists'
+    - Candle with no 'timestamp' will no longer use current UTC time
 
+- Candle's no longer require a Timestamp
 - Reworked and optimised 'append' and added 'insert' and 'prepend'
 - Added IndicatorCollection as a way to better reference Indicator's
 - More options when setting indicator 'Source' - str | Indicator | NestedSource
 - Updated Sub/Managed Indicator's to generate name based of parents if none set
 - Added Specific reading typing for Indicators
+- Reworked CandlestickType to be faster and can generate larger set of Candlesticks
 
 ---